requirements.txt
Django
django-test-extensions
#django-debug-toolbar
python-memcached
south

